requirements.txt
django
djangorestframework
django-extensions
python-dotenv
mysql-connector-python